condor_tests: add run_test.pl to replace batch_test.pl
run_test.pl is meant to replace batch_test.pl for batlab
testing. batch_test.pl contains a lot of dead code and legacy code
which does not fit with the current testing framework.

run_test.pl's code is derived from batch_test.pl. Its code is
lean as it's only objective is to run a single test given that
the environment to run the test is setup by the caller.